requirements/develope.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
clarifai==2.6.1
configparser==3.7.4
cryptography==2.6.1
emoji==0.5.2
future==0.17.1
googleapis-common-protos==1.5.10
grpcio==1.20.1
idna==2.8
jsonschema==2.6.0
pkg-resources==0.0.0
protobuf==3.7.1
pycparser==2.19
PySocks==1.6.8
python-telegram-bot==11.1.0
requests==2.21.0
six==1.12.0
urllib3==1.24.3
dnspython==1.16.0
pymongo==3.9.0